Remove old registries, fix NPE(hopefully)...
... and remove a bunch of useless documentation.
@k2b6s9j Javadoc(and scaladoc) are used so that people understand how to
use your functions. They should only be put on helper methods that are
meant to be available to everyone and on API classes and methods. And
DEFINETELY not on stuff like private fields :P Sure, comments can be
used in your own code to make sure you don't get lost, but look at this
mod! It's so simple! There are no big algorithms or compley trees of
functions. Everything is intuitive.